+++ data-raw/raw_wq.R
@@ -49,15 +49,6 @@ df_wq_2021_fmwt <-
 # Import EMP station coordinates from EDI
 df_coord_emp <- read_csv("https://portal.edirepository.org/nis/dataviewer?packageid=edi.458.4&entityid=827aa171ecae79731cc50ae0e590e5af")
 
-# Import region assignments
-df_regions <- read_csv("data-raw/Rosies_regions.csv")
-
-# Load Delta regions shapefile from Brian
-sf_delta <- R_EDSM_Subregions_Mahardja_FLOAT %>%
-  # Filter to regions of interest
-  filter(SubRegion %in% unique(df_regions$SubRegion)) %>%
-  select(SubRegion)
-
 
 # 2. Clean and Combine Data -----------------------------------------------
 
@@ -188,9 +179,9 @@ df_wq_all_c <- df_wq_all %>%
   # Convert to sf object
   st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326, remove = FALSE) %>%
   # Change to crs of sf_delta
-  st_transform(crs = st_crs(sf_delta)) %>%
+  st_transform(crs = st_crs(DroughtData:::sf_delta)) %>%
   # Add subregions
-  st_join(sf_delta, join = st_intersects) %>%
+  st_join(DroughtData:::sf_delta, join = st_intersects) %>%
   # Remove any data outside our subregions of interest
   filter(!is.na(SubRegion)) %>%
   # Drop sf geometry column since it's no longer needed
@@ -240,7 +231,7 @@ raw_wq_1975_2021 <- df_wq_all_c1 %>%
   # Restrict data to 1975-2021
   filter(YearAdj %in% 1975:2021) %>%
   # Add region designations
-  left_join(df_regions, by = c("Season", "SubRegion")) %>%
+  left_join(DroughtData:::df_regions, by = c("Season", "SubRegion")) %>%
   # Remove any data not chosen for the long-term analysis
   filter(Long_term) %>%
   # Select variables to keep